+++ hittables/box.h
@@ -23,7 +23,7 @@ public:
         sides.add(make_shared<flip_face>(make_shared<yz_rect>(p0.y(), p1.y(), p0.z(), p1.z(), p0.x(), material))); // right
     }
 
-    virtual bool hit(const ray& r, const float& t_min, const float& t_max, hit_record& rec) const override {
+    virtual bool hit(const ray& r, const double& t_min, const double& t_max, hit_record& rec) const override {
         return sides.hit(r, t_min, t_max, rec);
     }
