+++ hittables/rotate.h
@@ -13,7 +13,7 @@ class rotate_y : public hittable {
 public:
     rotate_y(std::shared_ptr<hittable> p, double angle);
 
-    virtual bool hit(const ray& r, const float& t_min, const float& t_max, hit_record& rec) const override;
+    virtual bool hit(const ray& r, const double& t_min, const double& t_max, hit_record& rec) const override;
     virtual bool bounding_box(const double& time0, const double& time1, aabb& output_box) const override {
         output_box = bbox;
         return has_box;
