+++ hittables/sphere.h
@@ -14,7 +14,7 @@ public:
     sphere() : radius(1) {}
     sphere(const vec3& cen, double r, std::shared_ptr<material> m) : center(cen), radius(r), mat_ptr(m) {};
 
-    virtual bool hit(const ray& r, const float& t_min, const float& t_max, hit_record& rec) const override;
+    virtual bool hit(const ray& r, const double& t_min, const double& t_max, hit_record& rec) const override;
     virtual bool bounding_box(const double& time0, const double& time1, aabb& output_box) const override;
 
     vec3 center;
